+++ docs/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Trigger Jira tracker syunc when Red Hat's CVSS change (OSIDB-4186)
 
 ## [4.12.0] - 2025-06-18
 ### Added